separate data models and API
